Add Docker development environment.
This is basically Debian with lots of packages installed for targeting
different architectures, and compilers.  It doesn't replace the
Dockerfiles in the test subdirectory since those give us a way to
target things Debian testing doesn't (for example, older compilers like
clang 3.8).

One of the big advantages of this is that it mounts the SIMDe source
directory inside of the container, so any changes in the host will
be visible in the container, which makes it possible to continue to
use the development environment in the host and just compile in the
container without requiring recompilation every time.